Add distcc support
By default, stage2 will run with distcc client enabled, with
DISTCC_HOSTS set to the original stage1 machine.  It will generate a
"distccd.script" to run on the stage1 host that uses the stage1 cross
compilers and a matching port configuration (not the default port) with
an unmodified distccd.  This script is to be manually run by whoever is
running stage2.  Some editing of distccd.script and the stage2
local.conf may be needed.